Move computePoints from GraphModel to PlottedFunctionAdornmentModel.
